backend/privylock_backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.0
cryptography==41.0.7
celery==5.3.4
redis==5.0.1
boto3==1.34.22
gunicorn==21.2.0

# NEW: Environment variables
django-environ>=0.10.0

# NEW: Google OAuth
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

